+++ internal/domain/discount_rules/cash_bulk_discount_rule.go
@@ -13,7 +13,7 @@ func (t CashBulkDiscountRule) Name() string {
 	return CashBulkDiscountRuleName
 }
 
-func (t CashBulkDiscountRule) TotalDiscount(quantity int64, _ float64) float64 {
+func (t CashBulkDiscountRule) TotalDiscount(quantity int, _ float64) float64 {
 	if quantity >= MinimumCashBulkSize {
 		return DiscountPerItem * float64(quantity)
 	}
