+++ internal/domain/shopping_cart_item.go
@@ -2,5 +2,5 @@ package domain
 
 type ShoppingCartItem struct {
 	Product  Product
-	Quantity int64
+	Quantity int
 }
